re orginize code
