+++ rl/network.py
@@ -1,0 +1,36 @@
+import torch.nn as nn
+import torch
+
+
+# create Policy Network
+class PolicyNet(nn.Module):
+    def __init__(self, state_dim, action_dim, hidden=64):
+        super().__init__()
+        self.net = nn.Sequential(
+            nn.Linear(state_dim, hidden),
+            nn.ReLU(),
+            nn.Linear(hidden, hidden),
+            nn.ReLU()
+        )
+        self.mean_head = nn.Linear(hidden, action_dim)
+        self.log_std = nn.Parameter(torch.full((action_dim,), -0.5))
+
+    def forward(self, state):
+        x = self.net(state)
+        mean = torch.sigmoid(self.mean_head(x))
+        std = torch.exp(self.log_std)
+        return mean, std
+
+
+# create value net for baseline
+class ValueNet(nn.Module):
+    def __init__(self, state_dim, hidden=64):
+        super().__init__()
+        self.value = nn.Sequential(
+            nn.Linear(state_dim, hidden),
+            nn.ReLU(),
+            nn.Linear(hidden, 1)
+        )
+
+    def forward(self, state):
+        return self.value(state).squeeze(-1)
